Tests and type error fixes
