+++ webpack/front_page/index.tsx
@@ -2,7 +2,8 @@ import { detectLanguage } from "../i18n";
 import { FrontPage } from "./front_page";
 import * as i18next from "i18next";
 import "../css/_index.scss";
-import { stopIE, attachToRoot } from "../util";
+import { attachToRoot } from "../util";
+import { stopIE } from "../util/stop_ie";
 
 stopIE();
