+++ webpack/util/__tests__/page_test.ts
@@ -1,4 +1,4 @@
-import { stopIE } from "../page";
+import { stopIE } from "../stop_ie";
 
 describe("stopIE()", () => {
   it("not IE", () => {
